Revert "tests/driver_sx127x: use uint8_t for NETOPT_RX_SYMBOL_TIMEOUT"
This reverts commit 5ef4c0a778ab7d4f625d63fdafe5e8347bfe479d.

+++ tests/driver_sx127x/main.c
@@ -359,7 +359,7 @@ int rx_timeout_cmd(int argc, char **argv)
     }
 
     netdev_t *netdev = &sx127x.netdev;
-    uint8_t rx_timeout;
+    uint16_t rx_timeout;
 
     if (strstr(argv[1], "set") != NULL) {
         if (argc < 3) {
